pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.9.2",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.2",
    "vecs>=0.4.5",
    "groq>=1.1.2",
    "pdfminer-six>=20260107",
    "fastembed>=0.8.0",
]
